Avalon

+++ Avalon/Demo/Entry.cs
@@ -2,7 +2,7 @@ namespace Demo;
 
 class Entry : EntryEntry
 {
-    protected override int ExecuteMain()
+    protected override long ExecuteMain()
     {
         Demo demo;
         demo = new Demo();
